refactor(#103): handle queue+open server-side
